pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Felipe Cabral", email = "felipe.benites@gmail.com"},
]
dependencies = [
    "langchain>=0.0.270",
    "PyPDF2",
    "python-dotenv",
    "streamlit",
    "openai",
    "faiss-cpu",
    "altair",
    "tiktoken",
    "huggingface-hub",
    "InstructorEmbedding>=1.0.1",
    "sentence-transformers>=2.2.2",
    "datetime>=5.2",
    "selenium>=4.11.2",
    "pdfminer-six>=20221105",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
start = "streamlit run app.py"